change strategy from "use as-is" to "please fork"
This repo was initially designed for anyone to be able to pick up and use.
As I add more and more of my own settings, I think this is pretty unrealistic.
Therefore, I've changed the README to suggest that users clone this repository
for themselves, and commit their information there. For this reason, my config
files will now be more personalised, and the `*_private` config system has been
removed.